Moved all `continuous_applications` specs
A few specs were overwritten with little to no change
