+++ spec/requests/candidate_interface/choose_course_spec.rb
