+++ spec/system/candidate_interface/course_selection/candidate_arrives_from_find_and_says_no_to_course_choice_spec.rb
