+++ spec/system/candidate_interface/course_selection/candidate_arrives_from_find_not_signed_in_spec.rb
