+++ spec/system/candidate_interface/course_selection/candidate_arrives_from_find_with_a_course_only_open_on_apply_spec.rb
