+++ spec/system/candidate_interface/course_selection/candidate_arrives_from_find_with_a_course_with_multiple_study_modes_spec.rb
