+++ spec/system/candidate_interface/course_selection/candidate_arrives_from_find_with_an_accepted_offer_spec.rb
