+++ spec/system/candidate_interface/course_selection/candidate_arrives_from_find_with_duplicate_application_spec.rb
