+++ spec/system/candidate_interface/course_selection/candidate_arrives_from_find_with_max_applications_spec.rb
