+++ spec/system/candidate_interface/course_selection/candidate_deletes_application_spec.rb
