+++ spec/system/candidate_interface/course_selection/candidate_edits_course_choices_spec.rb
