+++ spec/system/candidate_interface/course_selection/candidate_reapplying_to_a_course_spec.rb
