+++ spec/system/candidate_interface/course_selection/candidate_selecting_a_course_spec.rb
