+++ spec/system/candidate_interface/course_selection/candidate_selecting_a_course_study_mode_spec.rb
