+++ spec/system/candidate_interface/course_selection/candidate_selecting_a_course_with_multiple_sites_spec.rb
