+++ spec/system/candidate_interface/course_selection/candidate_selecting_a_duplicate_course_site_spec.rb
