+++ spec/system/candidate_interface/course_selection/candidate_visiting_selection_flow_for_selected_course_spec.rb
