+++ spec/system/candidate_interface/entering_details/candidate_marking_section_complete_or_incomplete_spec.rb
