+++ spec/system/candidate_interface/rejections/candidate_receives_rejection_email_spec.rb
