+++ spec/system/candidate_interface/submitting/candidate_has_an_application_become_inactive_spec.rb
