+++ spec/system/candidate_interface/submitting/candidate_is_redirected_away_from_complete_page_spec.rb
