+++ spec/system/candidate_interface/submitting/candidate_tries_to_submit_an_application_for_a_course_with_closed_application_spec.rb
