+++ spec/system/candidate_interface/submitting/candidate_tries_to_submit_an_application_for_course_unavailable_spec.rb
