+++ spec/system/candidate_interface/viewing-application-choices/candidate_sees_all_their_applications_spec.rb
