+++ spec/system/candidate_interface/viewing-application-choices/candidate_views_personal_statement_js_disabled_spec.rb
